Fix loading the preferences page setting the `currentonly` option to something truthy
I think the purpose of this function is normalize the value of
`currentonly` to exactly one of `true` or `false` in the most
reasonable manner possible in the face of whatever madness might have
been found in storage.

What one considers "reasonable" might vary from person to person, but I
propose the following:

If storage contains `null`, this is probably a first run, so default to
`true`.  If storage contains any other value, coerce the stored value
into a boolean value and use that.

Here's a table of some possible inputs that one might encounter, and the
outputs produced by the old and new code:

    input  | original output | new output | comments
    -------|-----------------|------------|--------------------------------------
    null   | true            | true       | normal (like a first run), no change
    true   | true            | true       | normal, no change
    false  | true            | false      | normal, BUG, changed
    "ASDF" | "ASDF"          | true       | pathological, normalize truthy value to true
    1      | 1               | true       | pathological, normalize truthy value to true
    0      | true            | false      | pathological, BUG, normalize falsy value to false